First steps towards Airable integration.
We add D-Bus into the mix.

This part of the API is going to be a good test for asynchronous
notifications of API clients. Whenever the login status changes, the
de.tahifi.Airable.ExternalServiceLoginStatus signal is sent around the
system and should be forwarded to clients. Implementing this part is
the next step of our Python/Werkzeug/Halogen journey.

Change-Id: Ia9079ed6ee60ac9d5b7f549f4767f13e076878c0