Move most initialization from global to app; set monitor port.
The monitor port is set based on the server port now: it is the server
port plus one. The monitor object is created during global
initialization, but the monitor thread is started only on first
invocation of our API entry point. This way, clients are forced to
enter through our entry point.

Change-Id: I2b687b404cfbea6a1b992e15800292904fb15fa6