airable: GET and PUT external music services credentials.
The PUT method also logs into or out from the service whose
credentials have been updated (not because it is a good idea to add
such kind of automatisms here (it is *not*), but because it is in
accordance with how Airable would like us to operate).

At this point, our API has reached a level where it can fully replace
our custom AppLink protocol.

Change-Id: I325702a4a4ebb593b9201b49269948a0c9354833