Add lots of documentation; fix typos; fix naming.
A few minor things had to be changed/cleaned up so that Sphinx could
process the code (it actually loads the code into an interpreter to
extract the documentation via introspection, so we cannot do too much
automatic stuff in __init__.py, and we must be a bit careful with what
we are doing in class definitions.

Let's finally gear this thing towards production grade quality.

Change-Id: I5eecaea92fdd8ec756553a66da0cfe95072cd41d