Endpoints are created with Link Relation Type.
For the sake of usability and readability, the optional parameters for
the Endpoint constructor are named parameters now.

Schema fields for embedded resources have been adapted to match the
Link Relation Types of the resources. Keeping these two consistent is
manual work, and their consistency cannot be verified automatically at
the moment.

Change-Id: I7fd04faf10b3f8210dce858b5223e861373d4d38