airable: Do not handle HTTP Cache-Control: header.
Our API must not act as a caching proxy or web cache. It would be a
big mistake to pull caching into our implementation because there is
software dedicated for caching web accesses, and we would almost
certainly interfere with it (not to mention the extra effort on our
side to get the caching right).

We could either put such a proxy on our Streaming Board (maybe
lighttpd supports this already?), or insist that clients must do their
own caching (it would be much easier there), but we must *not* do it
here, in the API implementation.

What the API should do instead, however, is adding support for ETag:
and If-None-Match: headers so that a cache can do the right thing.

Change-Id: Iaa73a6ff1f44fcc3c537e7d4e420814fd30af7a5