Move all embedded Schema class definitions to top level.
There used to be classes named Service.Schema or Status.Schema, so it
was not easy to do a meaningful search for a specific schema (full
qualification with dots required, search-word-under-cursor feature in
Vim didn't work to well). Problems would arise when importing a Schema
into some namespace, further complicating searches for uses of a
specific schema.

This commit was induced by a strange inconsistency in
strbo.airable.Credentials.SchemaShort, which was actually used to
serialize part of a strbo.airable.Service. Tracking down the problem
was annoying in part due to the way schemas were organized, so I've
fixed the inconsistency and established the new convention on schema
definitions.

This commit introduces a naming convention for schemas which
incorporates the name of the class it is meant to serialize. It also
pulls all schema definitions out of "their" classes (schemas are
actually not technically bound to classes because they rely on duck
typing) so that they can be moved around freely now.

Change-Id: Idf64474e409764d987a68ada52e3612644d61e82