network: Forward flag about caching of NICs and services.
In access point mode, Connman doesn't known about any WLAN service
anymore. For this reason, DCPD must cache information about services
known just before activating the AP. These cached information are
passed to the us when we ask for them.

It may be interesting for clients to know whether or not the service
or NIC data are cached (thus, possibly severely outdated), so we
forward the corresponding flags.

Change-Id: Ie49c8a948cdffe84e0f8a730279c3afbbe5d8714