network: Simplify WLAN configuration over service EP.
The configuration object sent to a service endpoint is not required to
contain a "wlan_info" field anymore; i.e., WLAN network information
are derived from the known service information. It was documented to
work this way, but the implementation was lacking...

Of course, configuration objects sent to interface endpoints still
need the "wlan_info" field.

Change-Id: I13d7e6c5c4fbd87705e02a75fafef4563d3abc21