recovery: Fix for gpg2 in StrBo V2.
There is gpg-agent now creating and deleting files, and we don't have
the opkg keyring anymore. To keep things smooth, gpg configuration
data are now stored below /var/local/etc/ so that gpg-agent can do
what it needs to do. Note that the home directory of strbo-rest is
used by lighttpd, so that's why the configuration data are put into a
completely different place.

Change-Id: I49061c731f3cdb23e1fd5d3e7456d7db424acfcb