endpoints: Turn uncaught exceptions into proper responses.
Any exception thrown by Werkzeug is turned into a response using the
get_response() method.

Other exceptions are treated depending on the debug mode. In case of
production mode, exceptions are turned into status 500 responses.
Information about the original exception, including a stack trace, are
passed back to the client as a JSON object so that we can debug them
when showing up in a log file. In case of debug mode, exceptions are
simply re-thrown so that the built-in debugger can handle them.

Change-Id: Ieb9e83ac6e0ff120d2b04dc11e5900331499c865