utils: Add function remove_file().
Actually we'd just like to use pathlib.Path.unlink(), but in case the
file in question does not exist, it throws an exception by default. In
Python 3.8 it is possible to avoid this by passing a flag, but for the
time being we need to support Python 3.6 as well (Ubuntu 18.04).

Our new remove_file() function avoids throwing in Python versions 3.6
and 3.8.

Change-Id: I9ee41f23b8295321aab436fcf95bf02d3fd71f18