system: Fix system update once again.
The UpdaTA script now operates on a well-defined state machine which
makes error handling and introspection much simpler. Our own helper
around that script creates a pid file when the script has been
launched in the background so that we can tell if the script is still
running.

A reboot loop still present in previous versions has been fixed as
well, triggered on successful update.

Change-Id: I3c1207321a9174d31fed5f6e895c0c273de44f20