helpers: Enable recovery data replacement after format.
Obviously, the images directory won't exist on a freshly formatted
recovery data partition. Our helper script, however, assumes the
directory exists and considers it an error if it doesn't.

So, the helper tries to create the directory now if it is not there.
This makes the script a little more fault tolerant, i.e., other
problems concerning mounting might go unnoticed after this change, but
so be it.

Change-Id: I21a29b656895497f9c172b2633ffce11f81e5117