player: Expose streamplayer D-Bus commands via REST API.
This is a direct connection to streamplayer, by-passing audio source
selection and other mechanisms. Also, these interfaces are not as
simple to use as they should be from a client's point of view, but
it's a start.

It's actually part of a bigger solution which also takes multi-client
action into account. The commands added in this commit represent part
of the set of interfaces a REST API user would use to control the
player while that user is holding the role of an active actor. What's
missing to complete the bigger solution are actor selection and
negotiation and player-agnostic request interfaces for passive actors
(request play, request stop, request next, request volume change,
etc.)

Change-Id: I4a60f4c2e04209ebb4879d5a8a06bdfd730ca583