monitor: Allow events for specific clients.
To make this work, the client needs to register itself with its
assigned client ID over its WebSocket connection. This allows the
event monitor to associate the socket connection with the client ID.

To register, the clients needs to send an object like this:

    {
        "op": "register",
        "client_id": 12345
    }

Events can thus be sent to one specific client with a given client ID,
and messages sent by the client could be augmented with the client ID
so that the receiver of the message could make use of it ("could"
because there is no use case for this yet).

Sending targeted events is done by setting the target_client_id
parameter for send_event(), send_endpoint(), send_error_object(), or
send_object().

Change-Id: Ibeef6e32265c04be8cf94b2de1ae9965a9e36d95