displays: Expose known displays in the system, allow sending content.
This commit is about the REST API design, not about the complete
representation of the whole system (this isn't possible at the moment
because the protocol to the ARM would need some extensions). The
implementation assumes there is always one display, it is built into
the device named "self", and it has 2 lines for displaying text.

Nevertheless, this implementation works already.

Change-Id: Ifbf0f1e808779ca0a83f3a7d5929d29a2b4bd36d