player_meta: Make use of the display.
The object returned by GET contains a link to the primary display now
so that the REST API client can simply POST its screen updates there.
It requires a valid secret key to do this.

In addition, the initial screen content can be set on audio source
activation to avoid glitches on the display.

Change-Id: Ided878608204bfd6f8613953ef5ed4eac33bc356