player_queue: Pushed meta data must be a proper JSON object.
The meta data were passed as a raw JSON string embedded into the push
request object, which was a bug in our app. This string had to be
parsed, and so we did, just to temporarily work around this bug and go
on with development. This bug is going to be fixed now, so we expect a
proper JSON object.

Change-Id: Ie2de9fb61a5e89e411410afee0d142229720d5ca