roon: Integrate taroon player.
All player status updates are forwarded to REST API clients now. Note
that control via /player/meta/requests was working already before.

The new RoonPlayerStatus class is heavily based on the previously
existing PlayerStatus class. In fact, these two could and should be
merged into a single class for DRY-ness. The Halogen schemas need to
be different, and that's all customization that should be required.

Change-Id: Iad4142095e9b29a9cfb971a5ce64b2fae278d15e