usb: Wrap call of pathlib is_file() into try/except block (#1912).
Seems like there is a Y2K38 problem in our Python version. It should
probably be recompiled with -D_FILE_OFFSET_BITS=64 or something like
that.

For the time being, we simply wrap the code which threw an @OSError@
75 (@EOVERFLOW@), "Value too large for defined data type", during
testing into try/except, and mark any entity for which pathlib throws
an error as "unknown" (instead of "file" or "dir").

Change-Id: I673e42dd9531071182fa23af474077e6735341f3